Rename assistantdb.py to skill_selection.py
